backend/requirements.txt
# Updated requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
ultralytics==8.0.223
opencv-python==4.8.1.78
google-generativeai==0.3.0
python-multipart==0.0.6
python-dotenv==1.0.0
sqlalchemy==2.0.23
alembic==1.13.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dateutil==2.8.2